[dicom_archive] Dicom archive siteproject main (#9549)
This adds advanced site and project permissions to the dicom archive.
It is designed to be fully backwards compatible based on the
`useImagingSiteProjectPermissions` configuration setting. Here is how
the permissions are designed to work

- If useImagingSiteProjectPermissions is disabled (status-quo)
  - A user with dicom_archive_allsites can see all the data (DEFAULT)
- A user with own sites can only see data at their own sites (can not
see data with no session ID)
- A user with own sites and nosessionid permission can see their site
data + dicoms not associated to a session
  - A user can always see DICOMS of scans they uploaded
- If useImagingSiteProjectPermissions is enabled
- A user with all sites permissions can see all the data from all sites
as long as the projectID is defined and they are affiliated with that
project (null sessions WILL NOT SHOW)
- A user with own sites, same logic, only projects they are affiliated
with (NO NULL sessions)
- A user with either site permissions + nosessionID permission can see
respectively the same as above + DICOMS not affiliated to a sessions
  - A user can always see DICOMS they have uploaded


TO BE NOTED: The core of this is the work done in the
`dicom_archive::getDataProvisionerWithFilters()` function. The code
there could have been made slightly more concise but I wrote it in that
way because I would like it to eventually become the DEFAULT
dataframework filtering logic since it accounts for most if not all
usecases (replacing the current UserSiteMatchOrHasAnyPermissions). It
defines flags that I would like to make standard going forward like
`dataSessionCanBeNull` where multiple module seem to lack a session
related to the resource.