[redcap] new module (#9474)
This PR adds the REDCap interoperability module.
This module opens an API endpoint to receive notifications from REDCap
servers through the use of the `Data Entry Trigger (DET)`. REDCap
notifications are parsed and REDCap records are imported into LORIS.
If an error happens during this notification handling process, an error
is generated in the issue tracker.
The module is imported from HBCD, and even if the core logic is the
same, REDCap objects are now their own models.
This version also supports connection to multiple REDCap instances.

Fixes #9473
Related PRs:
- #8181
- #8282