[dataquery] Return 0 results when a query is partially built (#9853)
When a query is being built there are situations where an empty set is
part of the query. This is logically equivalent to code that says "if (x
and) { ..}" and clearly wrong in the general case but can happen while
constructing the query as it's not yet complete and gets submitted to
get a count.

Update the query class to return 0 results in that case instead of
crashing.

Fixes #9851