The SQL engine now supports dictionary names containing question marks (#9883)
This fixes a bug that would cause the queries run by the SQL engine
to fail if any of the variables it depended on contained a question
mark. The PR escapes escapes the question marks so that PDO does not
interpret them as positional placeholders.

Reference: [https://www.php.net/manual/en/pdo.prepare.php](url)

Fixes #9884