Fix eslint line rules (#9796)
"npm run lintfix;js" currently generates hundreds of changes because the
rules enforced don't match the coding style of the codebase.

This fixes the rule for the number of lines between the tagblock and the
params to match what's actually used by the codebase, and fixes the
couple leftovers that were wrong.