[candidate_list] Translate candidate_list module (#9797)
This translates the candidate list module as an example of how to do
localization of a loris module that is written in React with class based
components.

JSON files are generated from the same `.po` files used by the PHP side
of file, which are then imported into the i18next resources. They are
generated at compile time with the i18next-conv package. Modules (in
this case candidate_list) must call the addResourceBundle method of
i18next to add the strings which are specific to their module.

This example uses the withTranslation HOC, but functional components can
use the useTranslation hook instead.