REDCap module improvements (#9839)
Small refactorings for REDCap module

List of changes:
- Extract REDCap / LORIS mapping methods from
`RedcapNotificationHandler` to a new `RedcapMapper` class, allowing to
use those methods without a REDCap notification, which is useful to
write a clean REDCap bulk import script.
- Move the `getIssueAssignee()` method from the `RedcapConfigParser`
class to the `RedcapQueries` (previously `Queries`) class, as the REDCap
issue assignee has previously been moved from the XML configuration to
the database.
- Correctly rename all relevant variables and methods to match REDCap
"event names" and "unique event names". In REDCap, an "event name" is
the user-facing name of an event, whereas a "unique event name" is the
internal identifier of an event used in most REDCap API routes. The
REDCap API naming is a little confusing on that matter. The differences
are now correctly reflected in our code.
- Rename some variables for better code expliciteness and consistency
(example: `$instrument` -> `$instrument_name`, snake_case as it is the
most used convention in the module).