[dqt - dataquery - permissions] Separate permissions for dqt and dataquery (#9918)
This creates a new permission (`dqt_view`), which is a duplicate of
`dataquery_view`; copying over current `perm_perm_action_rel` and
`user_perm_rel`. Its discriminator is the addition of ` (legacy)` to the
description seen in the `user_accounts` module.

A find-replace was performed on the `dqt` module to change all
references of the permission to `dqt_view`.

The `dataquery_view` permission should now be used for the `dataquery`
module.

Closes #9897