Remove confusing/broken logic from NDB_Client
The showPasswordExpiryScreen method in SinglePointLogin redirects
to a page that does not exist, and does it in a very hacky way.

This removes it to avoid confusion with the login workflow. Password
expiry (seem to be) currently handled by the login module on the frontend
while attempting to login, based on the response of the authenticate
endpoint.