[i18n] Dashboard -- My Tasks (#9750)
This builds on #9747 to add support for translating the "My Tasks"
widgets on the LORIS dashboard. In the current code, an "s" is hardcoded
to be added when count > 1, which doesn't work for other languages.

The signature had to be updated to include a singular and plural label,
as well as the text domain that the labels come from, since they come
from the module providing the widget and not the dashboard. (The
dngettext call also could not be done in the module, because the module
doesn't have the count to know if it's plural.)