[dashboard/intl] Add intl support for welcome message on dashboard
This adds localization to the "Welcome, (accountname)!" message on
the dashboard and the "Last login: " string. The last login date
is also now localized using the IntlDateFormatter to the user's locale
using the "traditional" date formatting.

The paragraph of text in the widget remains untranslated for the moment
because it comes from a database config setting.