[docs] Add intl dependency (#9991)
The IntlDateFormatter used in PR #9981 to provide locale specific date
formatting is a standard part of PHP, but requires the php8.3-intl to be
installed. Update the docs to mention it.

I believe we also require the curl extension in general now, because of
a dependency that uses it, so remove the note saying it's "for
development only".