[Localization] add LocalizableString class (#9997)
There are numerous places in the code where we need to provide a label,
but don't yet know the cardinality of the data that the label is going
to be provided for. (Often in widgets.)

This provides a "LocalizableString" class for strings that will
eventually need to be translated by dngettext, but where we don't yet
know the number of things for the label at the point where we are
creating it.