[issue_tracker] Null CenterID for new issues (#9974)
When opening the "Edit issue" page there was a 500 error because the
default CenterID for an issue was set to all of the user's sites. This
caused an array to string conversion error on line 265, but also does
not reflect what the default behaviour should be. By default, a blank
issue should have no site assigned.

* Resolves #9973