[dictionary] Fix override functionality in non-English languages (#10012)
The row name keys in the json passed by the DataTable are translated, so
the key looking up the "Field Name" needs to be translated when looking
up information from them.

Fixes #10011.

Also adds an explicit label for the "Cancel" button as it was not
translated (the label came from swal when not explicitly provided.)