[Multilingual] Simplify language imports
The importing of a named variable for every language at the top
of multilingual files, just to pass it to addResourceBundle is
getting a little unwieldy.

Replace them with an inline require() at the addResourceBundle call
site so that we don't need to pollute the import namespace for every
language (and to make life easier if dialects are added such as en-CA
and en-US).