[dashboard/multilingual] Add multilingual config settings (#9994)
This adds the backend infrastructure to allow multilingual config
settings with different values per language, so that pages such as the
project description on the dashboard can be localized.

ConfigSetting now has a "Multilingual" boolean flag to determine if the
config setting should be localizable. If it is multilingual, a new
`getMultilingualSetting` method on the NDB_Config object can be used to
get the value for a particular language code.

Translated strings are stored in a new "ConfigI18n" table (the casing
follows the standard of the Config tables, not the language table for
internal consistency.) When a translated version does not exist, the
`getMultilingualSetting` function will fall back to the non-localized
getSetting value.

There is not currently a frontend infrastructure to populate ConfigI18n.
The values must be done on the backend.