[Meta/Multilingual] Add wildcard target for .mo files
This adds a new wildcard target to convert .po files to .mo files.

Each `make modulename` target is updated to have a dependency on the
.mo files that it uses, rather than explicitly running msgfmt multiple times.

This means that when running `make modulename` it will only run the msgfmt
command if the po file has changed.

`make locales` should also be updated to make building loris more efficient
when doing development, but that it not done in this PR to avoid conflicts
with other module PRs that are in progress. It will be done ina future
PR.