[dashboard/statistics] Translate Recruitment widget (#10059)
This translates the recruitment widget (the last widget) on the LORIS
dashboard, so that the entire dashboard is now translated (The
recruitment widget had been waiting on #9937 to be merged before
translating it.). Japanese is used as a test language.

It also fixes a regression on the study progression translation
introduced by #9937 as new filters that are shared between the two were
added.