[candidate_profile/imaging_browser] Fix Imaging QC Summary widget on candiate_profile (#10050)
This fixes 2 problems which were preventing the Imaging QC summary
widget on the candidate_profile page from loading correctly:

1. The type of the VisitMap prop was incorrectly set to "array". It is
an object. This was causing a javascript error.
2. iterator_to_array was called after it had already been iterated on in
a foreach loop, resulting in the iterator_to_array call returning an
empty array. However, the result of the loop was not used anywhere, so
it is simply removed.