[user_accounts] Make examiner sites multiselect (#10056)
This modifies the user accounts module to have a multi-select field
for examiner sites instead of checkboxes. This way the full site names
are listed and a user can see more clearly what sites they are selecting
for an examiner. All other behaviour of examiner selection should remain
the same.

* Resolves #9864