[Core/Refactor] Do not mangle request URI (#10031)
The BaseRouter/Prefix router stripped off the handled part of the URI
before passing it to other handlers. The reasoning was that then the
modules could treat the request without regard to the location that
LORIS is being served for and act as if it's always the root.

However, in practice this causes more problems than it solves. The
original URI is lost completely, so middleware, error pages, and http
headers that may need it do not have access to it without reconstructing
it.

This updates the handled part of the URI to go into an "unhandledURI"
attribute while leaving the request->getURI untouched. Doing so allows
us to access the original URI when needed and if ie. you print
$request->getURI() in an error message you will get the correct value.