[Core] Add Support for TOTP / 2FA to LORIS (#9972)
This adds support for time based one time passwords (TOTP) to LORIS
based on RFC6238 (the standard used by authenticator apps such as Authy
or Microsoft Authenticator).

There is a new subpage of my_preferences where a user can register an
authenticator app to their account using a QR code. After validating the
code, the secret key used to generate it is saved to the users table.
After this point, any attempts to log in will check if they have a valid
2FA code by using a new MFA middleware. The middleware intercepts
requests and prompts for the code if it has not yet been provided.