[candidate_profile/media] Localize Candidate Media widget (#10077)
This translates the Candidate Media widget on the candidate_profile
page, and formats the date according to the locale.

"Last Modified" is changed to "Updated" to be consistent with widgets on
the dashboard and in order to make it easier to maintain translations.

"Last modified" does not appear to be used anywhere outside of the media
module.