[LINST] Fix loading of LINST instruments (#10079)
LINST instruments currently throw an exception because getExaminerNames
is called to get a list of examiners to pass to addElement. However,
getExaminerNames calls getField("Examiner") to get the default value.
Since addElement hasn't been called yet, it throws an exception saying
the field does not exist for the instrument.

This bypasses it by calling getInstanceData and getting the examiner key
directly.