[login] Add translations for login page (#10132)
This adds the ability to localize logged out pages and includes a
Japanese
translation for the login page and MFA pages.

If there are more than one language set up, there will be a dropdown in
the
page header allowing the user to select the language. The value is saved
in the PHP session. Upon login, the session cache of the selected
language is
unset so that the user's preference will take precedence.

en_CA is used as the default language.

The "title" and "StudyDescription" configs are now also multilingual
strings as they are used on the login page.