[issue_tracker] Add Japanese translations and fix issues (#10164)
This adds Japanese translations to the issue tracker and at the same
time fixes a number of issues:

1. the .pot file was not kept in sync with the hindi, so the strings
that had to be translated were not present. Added the missing ones to
the template file.
2. Date formatting is now done according to the user's locale in the
comment history
3. "Updated X seconds/minutes/hours ago" is now translated as a complete
sentence so that the grammar can be adjusted per language.
4. "Instrument" should be from the loris namespace
5. Translations for issues categories were in the po file, but the
markup to translate them in the code was not done, so they were not
getting translated.

Since the pot file was not up to date, it is possible that some strings
were missed.