[instruments] Add support for translation of instruments (#10181)
This adds support for the translation of instruments themselves in LORIS
(as opposed to the text around the instruments) building on the
multilingual support that is already present.

The instruments are translated according to the TimePoint's language,
rather than the logged in user's, as the page is used for data entry and
the screen should reflect the physical page.

Instrument metadata fields come from the "instruments" text domain,
while the question text / labels come from the text domain of the
instrument itself (ie. "mri_parameter_form" or "bmi"). The translations
should exist in the project's locale directory.

bmi and mri_parameter_form are provided in french for testing purposes
(mri_parameter_form is a PHP instrument, while the BMI it a LINST
instrument.)